Prime_bot/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
httpx>=0.27.0
chromadb>=0.5.0
sentence-transformers>=2.7.0
langchain-text-splitters>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
crewai>=0.41.1
crewai-tools>=0.4.26
pyyaml>=6.0.1
